+++ lib/main.dart
@@ -4,10 +4,12 @@ import 'package:flutter/services.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:zachranobed/features/appConfiguration/app_configuration.dart';
 import 'package:zachranobed/features/appConfiguration/mapper/app_configuration_mapper.dart';
+import 'package:zachranobed/features/foodboxes/di/food_box_dependency_container.dart';
 import 'package:zachranobed/features/login/di/login_dependency_container.dart';
+import 'package:zachranobed/features/offeredfood/di/offered_food_dependency_container.dart';
 import 'package:zachranobed/firebase/firebase_options.dart';
 import 'package:zachranobed/firebase/notifications.dart';
-import 'package:zachranobed/services/ioc_container.dart';
+import 'package:zachranobed/services/di/services_dependency_container.dart';
 import 'package:zachranobed/ui/widgets/app_root.dart';
 
 import 'common/firebase/firebase_helper.dart';
@@ -15,14 +17,16 @@ import 'common/firebase/firebase_helper.dart';
 void main() async {
   AppConfiguration.instance.set(
       AppConfigurationMapper.mapBuildConfiguration(appFlavor),
-      AppConfigurationMapper.mapApiConfiguration(appFlavor)
+      AppConfigurationMapper.mapApiConfiguration(appFlavor),
   );
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
 
   // DI Setup
-  IoCContainer.setup();
+  ServicesDependencyContainer.setup();
   LoginDependencyContainer.setup();
+  FoodBoxDependencyContainer.setup();
+  OfferedFoodDependencyContainer.setup();
 
   await Notifications().initNotifications();
 
@@ -33,7 +37,7 @@ void main() async {
   // Lock system preferences to portrait orientation only
   SystemChrome.setPreferredOrientations([
     DeviceOrientation.portraitUp,
-    DeviceOrientation.portraitDown
+    DeviceOrientation.portraitDown,
   ]);
 
   runApp(AppRoot());